cms-tools/pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=64"]

[project]
dependencies = ["ruamel.yaml"]
name = "cms-tools"
requires-python = ">= 3.12"
version = "0.1.0"

[project.scripts]
cms-tools = "cms_tools:main"

[tool.setuptools.package-data]
credentials = ["cms.conf", "conf.sample.yaml"]

[tool.pyright]
pythonPlatform = "All"
pythonVersion = "3.12"
typeCheckingMode = "strict"

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
ignore = [
    "ANN401",
    "D100",
    "D101",
    "D102",
    "D103",
    "D105",
    "D106",
    "D107",
    "D203",
    "D213",
    "E501",
    "E731",
    "PYI034",
]
select = [
    "ANN",
    "B",
    "C4",
    "COM",
    "D",
    "E",
    "F",
    "FA",
    "FBT",
    "FLY",
    "ISC",
    "N",
    "PERF",
    "PTH",
    "PYI",
    "RUF",
    "SIM",
    "UP",
    "W",
]
